Create /hello to get at postman; Added Student; Added /api/students; Added /api/students with @PathVariable; Added StudentErrorResponse and StudentNotFoundException; Added @ExceptionHandler; Added @ExceptionHandler for generics; Added DevTools in pom.xml; Refactory using @ControllerAdvice;
